+++ tests/test_requests_ep.py
@@ -94,15 +94,15 @@ class TestRequestsProps(TestRequests):
                 skip_non_strict=False,
             )
 
+        def test_django_filter_exclude(self):
+            pass
+
         def test_django_filter_multiple(self):
             self.do_test_django_filter(
                 reverse(f"{self.NAMESPACE}:enum-filter-multiple"),
                 multi=True,
                 skip_non_strict=False,
             )
-
-    else:
-        pass  # pragma: no cover
 
 
 TestRequests = None